Create Fluidteck build kits mini nasa type texh to teach young minds it is easy to do it with fluidteck as Lego even!!!
Fluidteck rockets explained further.
Health and safety is not just in work places as we do it just crossing the road as let’s not end up like a toad. Coz it is a crazy place out there with iver hyped older ones who use their distance of age as a way of either helping younger and older and same or being a bully and we know bullies are just stupid hahahaha! But they will love the logic as cant bully me not even god so don’t coz we are in the dream place but it is our reality here so let’s all have fun equally as it should be and play away all through the day and dream at night of what will and might.. everyday is a school day.